# docs/OpenAPIv3_swagger.yaml
components:
  schemas:
...
    domain.Export:
      properties:
        completed_at:
          type: string
        created_at:
          type: string
        error:
          type: string
        file_path:
          type: string
        format:
          $ref: '#/components/schemas/domain.ExportFormat'
        id:
          type: integer
        organization_id:
          type: integer
        status:
          $ref: '#/components/schemas/domain.ExportStatus'
        type:
          $ref: '#/components/schemas/domain.ExportType'
        updated_at:
          type: string
        user_id:
          type: integer
      required:
      - completed_at
      - created_at
      - error
      - file_path
      - format
      - id
      - organization_id
      - status
      - type
      - updated_at
      - user_id
      type: object
    domain.ExportFormat:
      type: string
      x-enum-varnames:
      - ExportFormatJSON
      - ExportFormatCSV
    domain.ExportStatus:
      type: string
      x-enum-varnames:
      - ExportStatusPending
      - ExportStatusProcessing
      - ExportStatusCompleted
      - ExportStatusFailed
    domain.ExportType:
      type: string
      x-enum-varnames:
      - ExportTypeChats
      - ExportTypeMessages
      - ExportTypeAll
...
    handler.ExportRequest:
      properties:
        format:
          enum:
          - json
          - csv
          type: string
        type:
          enum:
          - chats
          - messages
          - all
          type: string
      required:
      - format
      - type
      type: object
...
  securitySchemes:
    ApiKeyAuth:
      in: header
      name: x-organization-api-key
      type: apiKey
    BearerAuth:
      description: Type "Bearer" followed by a space and JWT token.
      in: header
      name: Authorization
      type: apiKey
...
info:
  contact:
    email: chatlogger-api-go+swagger@kjanat.com
    name: ChatLogger
    url: https://github.com/kjanat/chatlogger-api-go/issues
  description: API for logging and managing chat sessions.
  license:
    name: MIT License
    url: https://github.com/kjanat/chatlogger-api-go/blob/master/LICENSE
  termsOfService: https://github.com/kjanat/chatlogger-api-go#terms-of-service
  title: ChatLogger API (Go)
  version: ""
openapi: 3.1.0
paths:
...
  /v1/exports:
    get:
      description: Retrieves a paginated list of asynchronous export jobs for the
        user's organization.
      parameters:
      - description: Number of exports per page
        in: query
        name: limit
        schema:
          default: 10
          type: integer
      - description: Offset for pagination
        in: query
        name: offset
        schema:
          default: 0
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/domain.Export'
                type: array
          description: List of export jobs
        "401":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Unauthorized (JWT invalid/missing or Org ID not found)
        "500":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Failed to fetch exports
      security:
      - BearerAuth: []
      summary: List Export Jobs
      tags:
      - Exports
    post:
      description: Initiates an asynchronous job to export chat data (chats, messages,
        or all) in JSON or CSV format for the user's organization.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/handler.ExportRequest'
        description: Export parameters (format, type)
        required: true
      responses:
        "202":
          content:
            application/json:
              schema:
                additionalProperties: {}
                type: object
          description: 'export_id: uint64, status: domain.ExportStatus, message: string'
        "400":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Invalid request data (format/type)
        "401":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Unauthorized (JWT invalid/missing, Org or User ID not found)
        "500":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Failed to create export job
      security:
      - BearerAuth: []
      summary: Create Asynchronous Export Job
      tags:
      - Exports
  /v1/exports/{id}:
    get:
      description: Retrieves the status and details of a specific asynchronous export
        job.
      parameters:
      - description: Export Job ID
        in: path
        name: id
        required: true
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/domain.Export'
          description: Export job details
        "400":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Invalid export ID
        "401":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Unauthorized (JWT invalid/missing or Org ID not found)
        "404":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Export not found or doesn't belong to user's org
      security:
      - BearerAuth: []
      summary: Get Export Job Status
      tags:
      - Exports
  /v1/exports/{id}/download:
    get:
      description: Downloads the file generated by a completed asynchronous export
        job.
      parameters:
      - description: Export Job ID
        in: path
        name: id
        required: true
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                type: file
            application/octet-stream:
              schema:
                format: binary
                type: string
          description: Export file (JSON or CSV)
        "400":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Invalid export ID or export not ready
        "401":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Unauthorized (JWT invalid/missing or Org ID not found)
        "404":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Export not found or doesn't belong to user's org
        "500":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Export file path not found or file missing on disk
      security:
      - BearerAuth: []
      summary: Download Export File
      tags:
      - Exports
  /v1/exports/sync:
    post:
      description: Immediately generates and returns an export file (JSON or CSV)
        containing chat data. Use async export for large datasets.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/handler.ExportRequest'
        description: Export parameters (format, type)
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                type: file
            application/octet-stream:
              schema:
                format: binary
                type: string
          description: Export file (JSON or CSV)
        "400":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Invalid request data (format/type)
        "401":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Unauthorized (JWT invalid/missing or Org ID not found)
        "500":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Failed to retrieve data or generate export
      security:
      - BearerAuth: []
      summary: Create Synchronous Export (Legacy)
      tags:
      - Exports (Legacy)
...